requirements.txt
# DSV BizTalk to ACE Migration Tool - Requirements
# Core Dependencies for BizTalk to IBM ACE Migration Platform


groq>=0.4.1                    # Primary LLM client for Groq API
openai>=1.3.0                  # Alternative LLM support (if needed)
chromadb>=0.4.18               # Vector database for knowledge management
sentence-transformers>=2.2.2   # For text embeddings
langchain>=0.0.350             # LLM framework support
streamlit>=1.28.0              # Web application framework
PyPDF2>=3.0.1                 # PDF text extraction
pymupdf>=1.23.0               # Alternative PDF processing (fitz)
pdfplumber>=0.10.0            # Advanced PDF parsing
pandas>=2.1.0                 # Data manipulation and analysis
numpy>=1.24.0                 # Numerical computations
openpyxl>=3.1.0              # Excel file handling
lxml>=4.9.0                   # Advanced XML processing
xmltodict>=0.13.0             # XML to dictionary conversion
jsonschema>=4.17.0            # JSON schema validation
python-dotenv>=1.0.0          # Environment variable management
pydantic>=2.4.0               # Data validation
pathlib2>=2.3.7              # Enhanced path operations (if needed)
glob2>=0.7                    # Advanced file pattern matching
requests>=2.31.0              # HTTP client for API calls
httpx>=0.25.0                 # Async HTTP client
jinja2>=3.1.0                 # Template engine for code generation
black>=23.0.0                 # Code formatting (optional)
click>=8.1.0                  # Command line interface
rich>=13.0.0                  # Enhanced terminal output
tqdm>=4.66.0                  # Progress bars
python-dateutil>=2.8.0       # Date/time utilities
pytest>=7.4.0                # Testing framework
pytest-cov>=4.1.0            # Coverage reporting
mypy>=1.5.0                   # Type checking
plotly
psutil>=5.9.0                 # System monitoring
memory-profiler>=0.61.0       # Memory usage profiling
BeautifulSoup4                  # HTML parsing
typing-extensions>=4.7.0     # Extended typing support
setuptools>=68.0.0           # Build tools